rpk: refactor config command scopes
