+++ src/go/rpk/pkg/cli/cmd/cluster/config/config.go
@@ -12,10 +12,6 @@ package config
 import (
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
-	"github.com/vectorizedio/redpanda/src/go/rpk/pkg/cli/cmd/cluster/config/edit"
-	"github.com/vectorizedio/redpanda/src/go/rpk/pkg/cli/cmd/cluster/config/export"
-	"github.com/vectorizedio/redpanda/src/go/rpk/pkg/cli/cmd/cluster/config/importconfig"
-	"github.com/vectorizedio/redpanda/src/go/rpk/pkg/cli/cmd/cluster/config/status"
 	"github.com/vectorizedio/redpanda/src/go/rpk/pkg/cli/cmd/common"
 	"github.com/vectorizedio/redpanda/src/go/rpk/pkg/config"
 )
@@ -76,10 +72,10 @@ different redpanda version that does not recognize certain properties.`,
 	)
 
 	command.AddCommand(
-		importconfig.NewCommand(fs, &all),
-		export.NewCommand(fs, &all),
-		edit.NewCommand(fs, &all),
-		status.NewCommand(fs),
+		newImportCommand(fs, &all),
+		newExportCommand(fs, &all),
+		newEditCommand(fs, &all),
+		newStatusCommand(fs),
 	)
 
 	return command
